Added completed files
Added/uploaded completed versions of CalendarStuff.java, CalendarStuffTester.java, and CountTheDays.java for your enjoyment.

Remember there are LOTS of ways to do this -- nothing is *WRONG* as long as you get the correct answers for all the test cases.

Note that I also implemented the two methods that weren't specified as part of the assignment, and wrote specific tests for them, including testing for erroneous cases.  Feel free to use the test case file as a model for future work!